+++ server/models/pong.py
@@ -1,6 +1,7 @@
 from pydantic import BaseModel
 
 
+# 서버 상태를 확인할 때 사용하는 모델
 class Pong(BaseModel):
     status: bool
