--- server/static/sample.txt
@@ -1,1 +1,0 @@
-It's a sample text for /ping
